Determine working directory and media type.
If the working directory is not provided from the command line, then it is defaulted to the current working directory.

If the media type is not provided, then the user is prompted for it.

This commit also adds a simple menu system for getting a user selection from a list.