Implemented dry-run argument.
This allows the program to run through, but not actually perform any actions. This can be used on any unknown filename format to make sure the program will work as expected.

This commit also uncomments the rename command, so upon determining the season and episode number, the file will be moved to its ultimate destination.