Created actions dictionary to combine all move/downloads to one point.
In the first step of making a progress bar, all of the move commands are saved in a dictionary of lists. This allows all of the information to be determined for all files (like asking for the season/episode) and then performing all of the moves at once.

This also will allow us to know how many file moves will need to take place so we can track progress, and a smaller feature of asking for permission before actually executing the commands (in case something was entered wrong).